+++ src/internal/unix/zsh.js
@@ -3,21 +3,7 @@
  * @license MPL-2.0
  */
 
-/**
- * Escape an argument for use in Zsh.
- *
- * @param {string} arg The argument to escape.
- * @returns {string} The escaped argument.
- */
-function escapeArg(arg) {
-  return arg
-    .replace(/[\0\u0008\r\u001B\u009B]/gu, "")
-    .replace(/\n/gu, " ")
-    .replace(/\\/gu, "\\\\")
-    .replace(/(?<=^|\s)([#=~])/gu, "\\$1")
-    .replace(/(["$&'()*;<>?[\]`{|}])/gu, "\\$1")
-    .replace(/([\t ])/gu, "\\$1");
-}
+import RegExp from "../regexp.cjs";
 
 /**
  * Returns a function to escape arguments for use in Zsh for the given use case.
@@ -25,20 +11,35 @@ function escapeArg(arg) {
  * @returns {function(string): string} A function to escape arguments.
  */
 export function getEscapeFunction() {
-  return escapeArg;
+  const controls = new RegExp(/[\0\u0008\r\u001B\u009B]/g);
+  const newlines = new RegExp(/\n/g);
+  const backslashes = new RegExp(/\\/g);
+  const comments = new RegExp(/(^|\s)#/g);
+  const expansions = new RegExp(/(^|\s)([=~])/g);
+  const specials = new RegExp(/(["$&'()*;<>?[\]`{|}])/g);
+  const whitespace = new RegExp(/([\t ])/g);
+  return (arg) =>
+    arg
+      .replace(controls, "")
+      .replace(newlines, " ")
+      .replace(backslashes, "\\\\")
+      .replace(comments, "$1\\#")
+      .replace(expansions, "$1\\$2")
+      .replace(specials, "\\$1")
+      .replace(whitespace, "\\$1");
 }
 
 /**
  * Escape an argument for use in Zsh when the argument is being quoted.
  *
- * @param {string} arg The argument to escape.
- * @returns {string} The escaped argument.
+ * @returns {function(string): string} A function to escape arguments.
  */
-function escapeArgForQuoted(arg) {
-  return arg
-    .replace(/[\0\u0008\u001B\u009B]/gu, "")
-    .replace(/\r(?!\n)/gu, "")
-    .replace(/'/gu, "'\\''");
+function getQuoteEscapeFunction() {
+  const controls = new RegExp(/[\0\u0008\u001B\u009B]/g);
+  const crs = new RegExp(/(\r\n)|\r/g);
+  const quotes = new RegExp(/'/g);
+  return (arg) =>
+    arg.replace(controls, "").replace(crs, "$1").replace(quotes, "'\\''");
 }
 
 /**
@@ -57,5 +58,5 @@ function quoteArg(arg) {
  * @returns {(function(string): string)[]} A function pair to escape & quote arguments.
  */
 export function getQuoteFunction() {
-  return [escapeArgForQuoted, quoteArg];
+  return [getQuoteEscapeFunction(), quoteArg];
 }