+++ config/eslint.js
@@ -498,7 +498,7 @@ export default [
         "error",
         {
           ignoreSticky: false,
-          report: "potential",
+          report: "certain",
         },
       ],
       "regexp/no-trivially-nested-assertion": ["error"],