+++ src/internal/no-shell.js
@@ -1,9 +1,9 @@
 /**
- * @overview Provides functionality for shell-less escaping on Unix systems.
+ * @overview Provides functionality for shell-less escaping.
  * @license MPL-2.0
  */
 
-import RegExp from "../regexp.cjs";
+import RegExp from "./regexp.cjs";
 
 /**
  * The error message for use of quoting functionality.
@@ -14,7 +14,7 @@ import RegExp from "../regexp.cjs";
 const unsupportedError = "Quoting is not supported when no shell is used";
 
 /**
- * Returns a function to escape arguments for shell-less use.
+ * Returns a function to escape an argument for shell-less use.
  *
  * @returns {function(string): string} A function to escape arguments.
  */
@@ -25,7 +25,7 @@ export function getEscapeFunction() {
 }
 
 /**
- * Returns the provided value.
+ * Throws an error that this functionality is unsupported.
  *
  * @returns {never} Does not return.
  * @throws {Error} Always throws an error.