+++ test/compat/regexp-engine/unix.test.js
@@ -8,12 +8,18 @@ import * as bash from "../../../src/internal/unix/bash.js";
 import * as busybox from "../../../src/internal/unix/busybox.js";
 import * as csh from "../../../src/internal/unix/csh.js";
 import * as dash from "../../../src/internal/unix/dash.js";
-import * as nosh from "../../../src/internal/unix/no-shell.js";
 import * as zsh from "../../../src/internal/unix/zsh.js";
+import * as unix from "../../../src/internal/unix.js";
 
-const shells = [bash, busybox, csh, dash, nosh, zsh];
+const shells = [bash, busybox, csh, dash, zsh];
 
-const args = ["foobar", "Hello world!", "csh specific character: \u00A0"];
+const args = [
+  "foobar",
+  "Hello world!",
+  "csh specific character: \u00A0",
+  "--flag",
+  "-f",
+];
 
 export function testEscape() {
   for (const shell of shells) {
@@ -24,12 +30,15 @@ export function testEscape() {
   }
 }
 
+export function testFlagFunction() {
+  for (const arg of args) {
+    const flag = unix.getFlagFunction();
+    flag(arg);
+  }
+}
+
 export function testQuote() {
   for (const shell of shells) {
-    if (shell === nosh) {
-      continue;
-    }
-
     for (const arg of args) {
       const [escape, quote] = shell.getQuoteFunction();
       quote(escape(arg));