+++ test/unit/win/index.test.js
@@ -11,9 +11,9 @@ import * as fc from "fast-check";
 import * as ppTestKit from "pp-test-kit/simulate";
 import * as sinon from "sinon";
 
+import * as nosh from "../../../src/internal/no-shell.js";
 import { noShell } from "../../../src/internal/options.js";
 import * as cmd from "../../../src/internal/win/cmd.js";
-import * as nosh from "../../../src/internal/win/no-shell.js";
 import * as powershell from "../../../src/internal/win/powershell.js";
 import * as win from "../../../src/internal/win.js";