+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Versioning].
 
 - _No changes yet_
 
-## [2.1.12] - 2026-05-06
+## 2.1.12 (2026-05-06)
 
 - Add support for Node.js v26. ([#2520])
 - Make `--enable-experimental-regexp-engine` support opt-in. ([#2475])